--- base/templates/streams/what_we_do_block.html
@@ -1,90 +1,0 @@
-{% load wagtailimages_tags wagtailiconchooser_tags %}
-
-{% block extra_css %}
-
-<style>
-    .what-we-do-icon{
-        height: 60px;
-        width: 60px;
-        background-color: var(--primary-color-background);
-        border-radius: 50%;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-
-    .what-we-do-icon svg{
-        fill:  var(--primary-color);
-        height: 40%;
-        width: 40%;
-    }
-</style>
-
-{% endblock extra_css %}
-
-{% comment %} <div class="justify-content-center column {% if is_by3 %}is-one-third-tablet {% else %}is-one-quarter-desktop is-half-tablet {% endif %}is-full-mobile intro-item">
-    <figure class="image is-64x64">
-        <div class="what-we-do-icon">
-            {% if value.icon %}
-                {% svg_icon value.icon %}
-            {% endif %}
-        </div>
-    </figure>
-
-    <div style="display: flex;flex-direction: column;height: 100%;">
-        {% if value.title %}
-            <h5 class="what-we-do-item-title {% if is_by_3 %}is-by3 {% else %}is-by4{% endif %}">{{ value.title }}</h5>
-        {% endif %}
-        <p style="padding: 0">
-            {{ value.description }}
-        </p>
-    </div>
-</div> {% endcomment %}
-
-{% load wagtailcore_tags wagtailimages_tags lazyimages_tags %}
-
-<div class="accordion__item js-accordion-item active">
-    <h3 class="accordion-header js-accordion-header" style="display: flex;
-    align-items: center;
-    justify-content: space-between;">
-        
-        <div style="display:flex;align-items:center">
-
-            <figure class="image is-64x64 mr-4">
-                <div class="what-we-do-icon">
-                    {% if value.icon %}
-                        {% svg_icon value.icon %}
-                    {% endif %}
-                </div>
-            </figure>
-            <span>{{ value.title }}</span>
-        </div>
-        
-        <i class="fas fa-sort-down" style="float: right;"></i>
-    </h3>
-    <div class="accordion-body js-accordion-body"   style="display: none;">
-        <div class="accordion-body__contents">
-
-            <div class="columns is-multiline is-mobile " style="margin: 0;align-items: center">
-                {% if value.image %}
-                    <div class="column is-one-quarter-desktop is-full-mobile">
-                        <figure class="image" style="height: 180px">
-                            {% image value.image original as goal_img %}
-                            <img class="progressive__img progressive--not-loaded"
-                                 data-progressive="{{ goal_img.url }}"
-                                 src="{% lazy_image_url goal_img %}"
-                                 alt="{{ value.title }}"
-                                 style="height: 100%;width: 100%;object-fit: cover"
-                            />
-                        </figure>
-                    </div>
-                {% endif %}
-                <div class="column format-text">
-                    {{ value.description|richtext }}
-                </div>
-            </div>
-
-        </div>
-    </div>
-</div>
-
