--- nmhs_cms/templates/what_we_do_section.html
@@ -1,35 +1,0 @@
-{% load wagtailcore_tags i18n %}
-<div class="intro-items-container">
-    <div class="intro-items-header">
-        <span>
-            {% if what_we_do_title %}
-                {{ what_we_do_title }}
-            {% else %}
-                {% translate "What we do" %}
-            {% endif %}
-        </span>
-    </div>
-            <div class="columns is-mobile is-centered project-goals">
-                <div class="column is-three-quarters-desktop is-full-mobile">
-                    <div class="accordion js-accordion">
-                        {% for block in  page.what_we_do_items %}
-                            {% include_block block with is_first=forloop.first %}
-                        {% endfor %}
-                    </div>
-                </div>
-            </div>
-
-
-
-    {% if page.what_we_do_button_text and page.what_we_do_button_link %}
-        <div class="center">
-            <a href="{{ page.what_we_do_button_link.url }}"
-               class="button is-rounded center sec-btn cta-btn">
-                <span>{{ page.what_we_do_button_text }}</span>
-                <span class="icon">
-                    <i class="fas fa-arrow-right"></i>
-                </span>
-            </a>
-        </div>
-    {% endif %}
-</div>
